the_hero_rises/requirements.txt
Pillow == 6.2.1
attrs == 19.3.0
numpy == 1.17.4
opencv_python == 4.1.2.30
pathlib2 == 2.3.5
pycocotools == 2.0.0
pytorch_ignite == 0.2.1
torch == 1.3.1
torchvision == 0.4.2
trains == 0.13.1
tensorboard==2.1.0
